add 265 -- READ COMMIT
I was trying to do it as in 90 but easier as here it's all the
rectangles and not just the biggest. However, here sometimes rectangles
are allowed to overlap, EXCEPT ON ONE SINGLE TRAINING EXAMPLE (no 2). So
either there's an easier way of doing this that as a bug returns always
the expected output, or one has to hardcode that one example away, and
that's the reason of the weird restriction.

However this should be to optimize further, this is just a working
example (and through a hack so technically not even that)